Update to the latest dependencies. Add sample project for Micronaut. Make sure all sample projects run.
