+++ dagger/build.gradle
@@ -1,7 +1,8 @@
 dependencies {
-    compile 'com.squareup.dagger:dagger:1.2.2'
-    compile 'com.squareup.dagger:dagger-compiler:1.2.2'
+    compile 'com.squareup.dagger:dagger:1.2.5'
+    annotationProcessor 'com.squareup.dagger:dagger-compiler:1.2.5'
     compile project(':common')
+
 }
 
 ext.description = 'Creates a common abstraction over dagger'
