+++ guice/build.gradle
@@ -1,5 +1,5 @@
 dependencies {
-    compile 'com.google.inject:guice:4.0'
+    compile 'com.google.inject:guice:4.2.3'
     compile project(':common')
 }
