+++ settings.gradle
@@ -33,4 +33,3 @@ include 'dagger'
 include 'guice'
 include 'spring'
 include 'micronaut'
-
