+++ src/main/java/com/gluonhq/ignite/samples/ExampleApp.java
@@ -1,0 +1,10 @@
+package com.gluonhq.ignite.samples;
+
+import java.net.URL;
+
+public interface ExampleApp {
+
+    default URL getViewLocation() {
+        return getClass().getResource("/com/gluonhq/ignite/samples/View.fxml");
+    }
+}
