+++ src/main/resources/com/gluonhq/ignite/samples/View.fxml
@@ -34,7 +34,7 @@
 <?import javafx.scene.layout.*?>
 
 <BorderPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/8.0.40"
-            xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.gluonhq.ignite.ViewController">
+            xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.gluonhq.ignite.samples.ViewController">
    <top>
       <VBox minHeight="-Infinity" BorderPane.alignment="CENTER">
          <children>
